pyproject.toml
[project]
name = "senior-project"
version = "0.1.0"
description = "bulbul contests scoring system by deep learning"
readme = "README.md"
requires-python = ">=3.12.10"
dependencies = [
    "fastapi>=0.129.0",
    "librosa>=0.11.0",
    "matplotlib>=3.10.8",
    "motor>=3.7.1",
    "noisereduce>=3.0.3",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.22",
    "scikit-learn>=1.8.0",
    "scipy>=1.17.0",
    "seaborn>=0.13.2",
    "soundfile>=0.13.1",
    "tensorflow[and-cuda]>=2.20.0",
    "uvicorn>=0.41.0",
]
